Prepopulate recipients when replying to messages
When the 'reply' query parameter is not set, the fields are not
prepopulated even if the message is a reply. If it is set to any value
apart from 'all', reply to the sender is prepared. If it is set to
'all', reply to sender sender and all To/CC recipients is prepared.